susan-team-architect/backend/pyproject.toml
[project]
name = "susan-team-architect"
version = "0.1.0"
description = "Susan Team Architect — Python backend for AI agent orchestration and RAG"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.1.40",
    "voyageai>=0.3.0",
    "supabase>=2.20.0",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.10.0",
    "fastapi>=0.115.0",
    "PyYAML>=6.0",
    "firecrawl-py>=4.0.0",
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
